updated start.php to help reduce bundle setup
The idea here is to give the bundle user a hassle free installation and usage of the bundle.
The original bundle require customising the config/application.php file to add
the long lines of custom alias to link the namespaces of bootstrapper. instead,
with this commit, it will avoid that step for the bundle user. cheers :)