Allow Laravel 4 autopop when form bound to model
It's correct to pass an empty string rather than a null value to a textarea. But it's not necessary to do it here as the Laravel 4 Form::textarea() method will take care of this.

However, forcing it here is not just a duplication, it prevents Laravel 4's new model binding for forms feature from operating for textareas, as the Form::getValueAttribute() will receive an empty string rather than null and return that empty string as input instead of reading the value from the model.